Remove the site/apt/index.apt files that do not provide any useful information, or APT files that repeat the information existing in the package-info.java file.
Rational: those APT files were created in older time where the Maven generated site
was our main web site. This is no longer the case, and the Maven generated site is
now used only by a few developers. Since we should not put links from the public
site to the Jenkins build server (which host the Maven site), we should not have
any information for users in API files.


git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK7@1572147 13f79535-47bb-0310-9956-ffa450edef68