Allow NormalizedProjection constructors to know whether the second defining parameter of the Ellipsoid is the semi-major axis length or the inverse flattening factor, and in the later case allow constructors to get the definitive flattening factor in order to compute the excentricity more accuratly.
The intend is not to get more accurate map projection. Instead, we do that because some code sometime
compute back the original semi-minor axis length (for example) from the excentricity. Our goal is to
allow (if possible) such code to get a value closer to the original value (idealy the exact same value).


git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1693658 13f79535-47bb-0310-9956-ffa450edef68