Shapefile Features type, descriptor and database fields descriptors are avaiable in the Shapefile object by calling the methods :     public DefaultFeatureType getFeaturesType();     public ShapefileDescriptor getShapefileDescriptor();     public List<DBase3FieldDescriptor> getDatabaseFieldsDescriptors();
For information : 
new ShapeFile(String filename) 
doesn't do any I/O operation and is just a declaration of a ShapeFile than can be used later.
The descriptors becomes available as soon as a reading attempt is made by a call to the readFeature() method, or by a call of loadDescriptors() on the ShapeFile. The binary headers of the shapefile and dbf has to be readen at these times to get the descriptors.

git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1713851 13f79535-47bb-0310-9956-ffa450edef68