Renamed "resource" property as "data".
Not all kinds of resources are digital data. For example a resource may be an organization, or citation of facts, tables and figures printed on paper, photographic material, or other media (see all `org.opengis.metadata.citation.PresentationForm` values having the `_HARDCOPY` suffix in their name). The resource in a MapLayer should be a resource containing digital data.