When formatting projected coordinates, give also the axis direction. Example: "-100 m E 300 m N". Maybe a future version should replace "E" by "W" when the value is negative, but current version is enough for resolving the ambiguity problem.
Add an `setGroundAccuracy(Quantity)` method for specifying a text like "± 1 m" to append after coordinate values.
This accuracy is shown only if the precision of coordinate values (i.e. the number of fraction digits) is smaller
than the specified accuracy.

Those changes are applied only on the `CoordinateFormat.format(…)` method. The `CoordinateFormat.parse(…)` method
is not yet updated accordingly.