Change EnvelopeOperation.serialVersionUID since the meaning of null values in `attributeToCRS` changed. Since we broke serialization compatibility anyway, opportunistically rename `crs` as `targetCRS` for making clearer that this is not necessarily the CRS of geometry objects stored in properties.
Then change the approach about how to get the CRS. The previous approach (search early in attribute characteristics)
was designed before a CRS can be associated directly to a GeometryWrapper. But we more recent Apache SIS code state,
there is better chance that `Envelope.getCoordinateReferenceSystem()` returns a non-null value, so we should wait a
bit more before to invoke `Feature.getProperty(…)` instead of `Feature.getPropertyValue(…)`.