fix(Feature): Ensure coordinate order consistency when converting an envelope to a geometry
Previous behavior tried to build a geometry in a right-handed space. However, the coordinate reference system was not adapted to the potential axis swapping.

The strategy has been changed to preserve input coordinate order. The swapping is mostly needed for interoperability with third-party contexts/applications. As such, it should be managed in higher levels.