Change the way "North pole rotation" is created and change the way the inverse operation is created. The north pole case is now implemented as a "South pole rotation" operation rotating the antipodal point. It causes a 180° offset in longitude, but this is intended according the following definition from COSMO:
> It is convenient to define the rotated meridian which runs through both the geographical and the rotated North Pole as the 0◦ meridian.

In addition, the way to create inverse operation has been modified in order to keep latitude in the range of valid values.
Mathematically, it is sometime convenient to use an out-of-range value as a trick for avoiding the 180° offset in longitudes,
and sometime it is convenient to keep the offset. We select whatever method keep longitude outputs closer to [-180 … 180]° range.