Improvement in the cache of `RenderedImage` instances: - Revisit the `equals(Object)` and `hashCode()` methods. - Reuse existing `RenderedImage` instances for a given `SliceExtent`.
The latter item is helpful even if the rendered images are cheap to create
(some of them are just wrappers around existing `DataBuffer`) because the
image may be the starting point of a chain of operations with statistics,
raster resampling, etc. Not reusing an existing image means discarding
all the derivated information.