fix(Core): add control points if required when converting an envelope to a polygon.
On axes with potential wrap-around, if the envelope is very large, the conversion to a rectangle without additional control points create an ambiguity.
When trying to link two points of the rectangle using shortest path, the result could not follow rectangle edges, due to the wrap-around. Adding control points mitigate that problem.

Note: the current solution is only a quick fix, surely there is room for improvement. For example, the current solution only consider a single wrap-around period.