Review of the new style package added in previous commit. Summary of changes:
* Remove usages of GeoAPI-pending interfaces (see below).
* Warning resolutions for raw parametric types and unchecked casts.
* Add some properties defined by the SE standard but missing in the classes.
* Empty placeholders when addition of missing properties would be too much work.
* Use `java.util.Optional` for better identification of which properties are optional.
* Add some (not all) JAXB annotations, as a help for identifying mandatory properties.
* Lazy initialization of (sometime standardized) default values of mandatory properties.
* Add credits to OGC GeoAPI 2 contributors since GeoAPI was a source of inspiration.
* Set version number to 1.5 (we will not include this package in SIS 1.4 release).
* Documentation.

The most important change is the removal of usages of GeoAPI-pending interfaces.
The GeoAPI `org.opengis.style` package is not part of a formal GeoAPI release.
It is currently found only in the "GeoAPI-pending" module, which is not released.
Those interfaces were based on XML encoding as defined in OGC 05-077r4 standard.
Using XML schema instead of UML models has many drawbacks shown in OGC meetings.
UML abstract model was not available when `org.opengis.style` was first designed.
But today, the ISO 19117:2012 standard evolved with a more elaborated model.
There is also various OGC working groups actively working on new style standards,
with goals such as better representation of 3D data. For those reasons, the style
interfaces in GeoAPI-pending may change and it would premature to commit to them.

https://issues.apache.org/jira/browse/SIS-583