Rename `org.apache.sis.internal.style` package as `org.apache.sis.style.se1`. By moving the package outside `internal` namespace, we are making it public. However we still target SIS 1.5, not SIS 1.4, for release.
The `se1` name stands for "OGC Symbology Encoding 1".
We do not commit this package as a generic style API,
because there is revision in progress in OGC and ISO 19117.
We use SE1 as an API that we can use for now, but with the
intend to make a more generic API available in the future.