Modernize the check for	null argument value:
Some (but not all) calls to `ArgumentChecks.ensureNonNull` are replaced by `Objects.requireNonNull`,
because the latter has optimization (@ForceInline) that we cannot match. It may	also be	intrinsified
in a future Java version. However, we do this replacement only when there is few ambiguity about which
argument was null. We keep the old method when there is many arguments, or when the null check is done
in a non-obvious place.

Check for null value is completely removed when we can rely on the implicit null check, for example
when the argument is dereferenced in the first few lines. We can do that because since Java 14, the
NullPointerException messages are automatically generated (https://openjdk.org/jeps/358).
Many of the previous checks were workaround for the non-helpful error messages in older Java versions.

This commit contains also two opportunistic changes:
- `GenericName` constructors do not allow anymore empty names (`ensureNonNull` → `ensureNonEmpty`).
- Opportunistic revision of a few `@SuppressWarnings` annotations.