Deprecate for removal the `AbstractFeature.getValueOrFallback(…)` method because experience suggests that it encourages bugs in user's codes that stay unnoticed. That method can be replaced by the new `FeatureType.hasProperty(String)` method. However, that latter method is actually not needed often.
The most important replacement for `getValueOrFallback(…)` is to do a better handling
of `PropertyNotFoundException` as warnings. When that warning occurs, this is usually
because we have not done a good job in eliminating dead expressions and dead filters.
So instead of invoking `hasProperty(String)` everywhere, this commit improves filter
optimizations so that a `PropertyNotFoundException` would be a real error.